2023-06-08 main 브랜치 통합

+++ src/main/java/com/jxx/xuni/review/dto/request/ReviewForm.java
@@ -8,7 +8,7 @@ import org.hibernate.validator.constraints.Range;
 public record ReviewForm(
         @NotNull(message = "null 일 수 없습니다")
         @Range(min = 0, max = 5, message = "0 - 5 사이의 점수를 매겨주세요")
-        Byte rating,
+        Integer rating,
         @NotBlank(message = "공백일 수 없습니다")
         @Pattern(regexp = "^.{10,500}$", message = "10자 이상 작성해주세요. 500자 까지")
         String comment,
